smart-doc-platform/requirements.txt
fastapi
uvicorn
google-generativeai
python-docx
PyMuPDF
docx2pdf
python-dotenv
python-multipart
transformers
torch
sentencepiece
